+++ CHANGES.md
@@ -3,6 +3,8 @@
 #### next release (8.2.27)
 
 - Change icon used for display group remove button
+- Make access control UI compatible to Magda v1 and v2 with v2 overriding v1.
+- [The next improvement]
 
 #### 8.2.26 - 2023-03-21